refactor: Fixing intializer into single file
