--- initializers/connectToDB.go
@@ -1,24 +1,0 @@
-package initializers
-
-import (
-	"log"
-	"os"
-
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
-)
-
-var DB *gorm.DB
-
-func ConnectToDB() {
-
-	var err error
-
-	dsn := os.Getenv("DB")
-	DB, err = gorm.Open(mysql.Open(dsn), &gorm.Config{})
-
-	if err != nil {
-		log.Fatalln("Connection to database error")
-	}
-
-}
