--- initializers/loadEnvVariables.go
@@ -1,16 +1,0 @@
-package initializers
-
-import (
-	"log"
-
-	"github.com/joho/godotenv"
-)
-
-func LoadEnvVariables() {
-	err := godotenv.Load()
-
-	if err != nil {
-		log.Fatalln("Error loading .env file")
-	}
-
-}
