--- initializers/syncDatabase.go
@@ -1,9 +1,0 @@
-package initializers
-
-import (
-	"github.com/AvinFajarF/model"
-)
-
-func Migrate() {
-	DB.AutoMigrate(&model.User{}, &model.Task{})
-}
